add foundry-devOps to get recent deployments, add solmate for erc20 contracts, update foundry toml to reflect 1st two changes, update interactions and helperconfig to deploy link token, add more tests to RaffleTest

+++ src/Raffle.sol
@@ -59,7 +59,7 @@ contract Raffle is VRFConsumerBaseV2Plus {
     // @dev duration of the lottery in seconds
     uint256 private immutable i_interval;
     bytes32 private immutable i_keyHash;
-    uint64 private immutable i_subscriptionId;
+    uint256 private immutable i_subscriptionId;
     uint32 private immutable i_callbackGasLimit;
     address payable[] private s_players;
     uint256 private s_lastTimeStamp;
@@ -77,7 +77,7 @@ contract Raffle is VRFConsumerBaseV2Plus {
         uint256 interval,
         address vrfCoordinator,
         bytes32 gasLane,
-        uint64 subscriptionId,
+        uint256 subscriptionId,
         uint32 callbackGasLimit
     ) VRFConsumerBaseV2Plus(vrfCoordinator) {
         i_entranceFee = entranceFee;
